requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-flask==1.3.0
python-dotenv==1.0.1
